requirements.txt
azure-cognitiveservices-speech
azure-identity
flask
openai
pytz
requests
python-dotenv

